Remove one-stop-shop CIF ontology
The reason is that the version of the CIF top ontology cannot be
uniquely defined for all subsequent imports of CIF_*.ttl files.
Say, for example, `cif_core.dic` uses v1.0.0 of `cif_top.ttl` and
`cif_sym.ttl` uses v0.9.5 of `cif_top.ttl` (or similar), the `cif.ttl`
file cannot handle this and import both ontologies. Or rather, there is
a high possibility that these two versions cannot be accommodated at the
same time.